Address review comments

+++ dashboard/src/main/home/project-settings/BillingPage.tsx
@@ -1,5 +1,6 @@
-import React, { useContext, useEffect, useState } from "react";
+import React, { useContext, useState } from "react";
 import ParentSize from "@visx/responsive/lib/components/ParentSize";
+import { intlFormatDistance } from "date-fns";
 import styled from "styled-components";
 
 import Loading from "components/Loading";
@@ -12,8 +13,8 @@ import Spacer from "components/porter/Spacer";
 import Text from "components/porter/Text";
 import {
   checkIfProjectHasPayment,
-  useCustomerDashboard,
   useCustomerPlan,
+  useCustomeUsageDashboard,
   usePaymentMethods,
   usePorterCredits,
   useSetDefaultPaymentMethod,
@@ -44,48 +45,24 @@ function BillingPage(): JSX.Element {
 
   const { refetchPaymentEnabled } = checkIfProjectHasPayment();
 
-  const { url: usageDashboard } = useCustomerDashboard("usage");
+  const { url: usageDashboard } = useCustomeUsageDashboard("usage");
 
   const formatCredits = (credits: number): string => {
     return (credits / 100).toFixed(2);
   };
 
-  const monthDiff = (d1: Date, d2: Date) => {
-    var months;
-    months = (d2.getFullYear() - d1.getFullYear()) * 12;
-    months -= d1.getMonth();
-    months += d2.getMonth();
-    return months <= 0 ? 0 : months;
-  };
-
-  const daysDiff = (d1: Date, d2: Date) => {
-    const _MS_PER_DAY = 1000 * 60 * 60 * 24;
-    const utc1 = Date.UTC(d1.getFullYear(), d1.getMonth(), d1.getDate());
-    const utc2 = Date.UTC(d2.getFullYear(), d2.getMonth(), d2.getDate());
-
-    return Math.floor((utc2 - utc1) / _MS_PER_DAY);
-  };
-
   const relativeTime = (timestampUTC: string): string => {
     const tsDate = new Date(timestampUTC);
     const now = new Date();
-
-    const remainingMonths = monthDiff(now, tsDate);
-    const remainingDays = daysDiff(now, tsDate);
-
-    const relativeFormat = remainingMonths > 0 ? "months" : "days";
-    const relativeValue = remainingMonths > 0 ? remainingMonths : remainingDays;
-
-    const rt = new Intl.RelativeTimeFormat("en", { style: "short" });
-    return rt.format(relativeValue, relativeFormat);
+    return intlFormatDistance(tsDate, now);
   };
 
-  const readableDate = (s: string) => new Date(s).toLocaleDateString();
+  const readableDate = (s: string): string => new Date(s).toLocaleDateString();
 
-  const onCreate = async () => {
-    await refetchPaymentMethods();
+  const onCreate = async (): Promise<void> => {
+    await refetchPaymentMethods({ throwOnError: false, cancelRefetch: false });
     setShouldCreate(false);
-    refetchPaymentEnabled();
+    await refetchPaymentEnabled({ throwOnError: false, cancelRefetch: false });
   };
 
   if (shouldCreate) {
@@ -131,25 +108,30 @@ function BillingPage(): JSX.Element {
                     <Container row={true}>
                       <DeleteButton
                         onClick={() => {
-                          setCurrentOverlay({
-                            message: `Are you sure you want to remove this payment method?`,
-                            onYes: () => {
-                              deletePaymentMethod(paymentMethod.id);
-                              setCurrentOverlay(null);
-                            },
-                            onNo: () => {
-                              setCurrentOverlay(null);
-                            },
-                          });
+                          if (setCurrentOverlay) {
+                            setCurrentOverlay({
+                              message: `Are you sure you want to remove this payment method?`,
+                              onYes: async () => {
+                                await deletePaymentMethod(paymentMethod.id);
+                                setCurrentOverlay(null);
+                              },
+                              onNo: () => {
+                                setCurrentOverlay(null);
+                              },
+                            });
+                          }
                         }}
                       >
                         <Icon src={trashIcon} height={"18px"} />
                       </DeleteButton>
                       <Spacer inline x={1} />
                       <Button
-                        onClick={() => {
-                          setDefaultPaymentMethod(paymentMethod.id);
-                          refetchPaymentMethods();
+                        onClick={async () => {
+                          await setDefaultPaymentMethod(paymentMethod.id);
+                          await refetchPaymentMethods({
+                            throwOnError: false,
+                            cancelRefetch: false,
+                          });
                         }}
                       >
                         Set as default
@@ -209,7 +191,7 @@ function BillingPage(): JSX.Element {
             </Text>
             <Spacer y={1} />
 
-            {plan != null && plan.plan_name != "" ? (
+            {plan !== undefined && plan.plan_name !== "" ? (
               <div>
                 <Text>Active Plan</Text>
                 <Spacer y={0.5} />
@@ -247,14 +229,13 @@ function BillingPage(): JSX.Element {
                         src={usageDashboard}
                         scrolling="no"
                         frameBorder={0}
-                        allowTransparency={true}
                       ></iframe>
                     )}
                   </ParentSize>
                 </Container>
               </div>
             ) : (
-              <Text>This project doesn't have an active billing plan.</Text>
+              <Text>This project does not have an active billing plan.</Text>
             )}
           </div>
         </div>
